moving cellmetris to analysis and parameters_for_isx to files_io
