+++ cgk_calcium_tools/analysis_utils.py
@@ -2,11 +2,95 @@ import isx
 import pandas as pd
 from pathlib import Path
 import numpy as np
+import os
+from .files_io import write_log_file, same_json_or_remove, parameters_for_isx
 
-def cellmetrics(fh, cellsetname="pca-ica"):
-    ev = isx.EventSet.read()
-    cs = isx.CellSet.read()
+def ixs_cell_metrics(fh, cellsetname: str, verbose=False) -> pd.DataFrame:
+    """
+    This function runs the isx.cell_metrics function, which compute cell metrics
+    for a given cell set and events combination
 
+    Parameters
+    ----------
+    fh : isx_files_handler
+        isx_files_handler object
+    cellsetname : str
+        cell label to get filename
+    verbose : bool, optional
+        Show additional messages, by default False
+
+    Returns
+    -------
+    pd.DataFrame
+        a concatenates list with metrics
+
+
+    """
+    cell_set_files = fh.get_results_filenames(
+        f"{cellsetname}", op=None, single_plane=False
+    )
+    ed_files = fh.get_results_filenames(
+        f"{cellsetname}-ED", op=None, single_plane=False
+    )
+    metrics_files = fh.get_results_filenames(
+        f"{cellsetname}_metrics.csv", op=None, single_plane=False
+    )
+
+    for cellset, ed, metric in zip(cell_set_files, ed_files, metrics_files):
+        inputs_args = {
+            "input_cell_set_files": os.path.basename(cellset),
+            "input_event_set_files": os.path.basename(ed),
+            "output_metrics_files": os.path.basename(metric),
+        }
+        if not same_json_or_remove(
+            inputs_args,
+            output=metric,
+            verbose=verbose,
+            input_files_keys=["input_cell_set_files", "input_event_set_files"],
+        ):
+            try:
+                isx.cell_metrics(
+                    **parameters_for_isx(
+                        inputs_args,
+                        to_update={
+                            "input_cell_set_files": cellset,
+                            "input_event_set_files": ed,
+                            "output_metrics_files": metric,
+                        },
+                    )
+                )
+            except Exception as e:
+                print(e)
+            write_log_file(
+                inputs_args,
+                os.path.dirname(metric),
+                {"function": "cell_metrics"},
+                input_files_keys=["input_cell_set_files", "input_event_set_files"],
+                output_file_key="output_metrics_files",
+            )
+
+    df = []
+    for metric_file, label, cell_set_file in zip(
+        metrics_files, fh.recording_labels, cell_set_files
+    ):
+        aux = pd.read_csv(metric_file)
+        aux["Recording Label"] = label
+        cell_set = isx.CellSet.read(cell_set_file)
+        num_cells = cell_set.num_cells
+        status = pd.DataFrame.from_dict(
+            {
+                cell: [cell_set.get_cell_name(cell), cell_set.get_cell_status(cell)]
+                for cell in range(num_cells)
+            },
+            orient="index",
+            columns=["cellName", "status"],
+        )
+        cell_set.flush()
+        aux = aux.merge(status, on="cellName")
+
+        df.append(aux)
+
+    return pd.concat(df)
 
 def compute_events_rates(events_df,cellsets_df):    
     fr = events_df.groupby(["Recording Label", "cell_name"])["Time (us)"].count().rename('#events').reset_index()
